+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Flutter 3.27 compatibility ( by @slavap )
 * Introduce TitleRenderer for customizable column titles ( by @doonfrs )
 * Implement custom loading widget support and add Loading Options feature to documentation and demo ( by @doonfrs )
+* Introduce Percentage Column Type ( by @doonfrs )
 
 ## [1.6.5] - 2025. 03. 22